backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.3
botocore==1.38.3
certifi==2025.1.31
click==8.1.8
dlib==19.24.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.5
passlib==1.7.4
Pillow==9.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==1.10.13
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
s3transfer==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
qrcode[pil]==7.4.2
watchfiles==1.0.5
websockets==15.0.1

